refactor: rename `navbar.tsx` into `Navbar.tsx`
