+++ app/(main)/layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from '@/components/navbar'
+import { Navbar } from '@/components/Navbar'
 
 export default function MainLayout({
   children,
