+++ app/not-found.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from '@/components/navbar'
+import { Navbar } from '@/components/Navbar'
 import { Button } from '@/components/ui/button'
 import { TriangleAlertIcon } from 'lucide-react'
 import type { Metadata } from 'next'
